backend/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.1
drf-writable-nested==0.7.0
gunicorn==23.0.0
psycopg2==2.9.9
python-dotenv==1.0.1
pre-commit==3.8.0
django-stubs==5.1.0
djangorestframework-stubs==3.15.1
drf-writable-nested==0.7.0
mailtrap==2.0.1
openai==1.58.1
django-redis==5.4.0
django-extensions==3.2.3
str2bool==1.1